requirements.txt
python-dotenv
pytest
langchain-community
unstructured 
faiss-cpu 
langchain-openai 
langchain-huggingface
langgraph
gradio